Generical EntityManager test class
